Cleanup some leftovers
