+++ .github/templates/pr.yml
@@ -25,7 +25,7 @@ jobs:
   build-wrappers:
     uses: ./.github/workflows/wrappers.yml
     name: Wrappers
-  _: #@ template.replace(deployBaas(["Code Coverage", "iOS"]))
+  _: #@ template.replace(deployBaas(["Code Coverage"]))
   build-packages:
     uses: ./.github/workflows/build-packages.yml
     name: Package
@@ -39,7 +39,7 @@ jobs:
   _: #@ template.replace(runTests("UWP Managed", runSyncTests = False, additionalSecrets = ["Pfx_Password", "Base64_Encoded_Pfx"]))
   _: #@ template.replace(runNetCoreTests('["net6.0"]'))
   _: #@ template.replace(runTests("macOS", runSyncTests = False))
-  _: #@ template.replace(runTests("iOS", runSyncTests = True))
+  _: #@ template.replace(runTests("iOS", runSyncTests = False))
   _: #@ template.replace(runTests("Android", runSyncTests = False, additionalSecrets=["AWS_DEVICEFARM_ACCESS_KEY_ID", "AWS_DEVICEFARM_SECRET_ACCESS_KEY", "DEVICEFARM_PROJECT_ARN", "DEVICEFARM_ANDROID_POOL_ARN"]))
   _: #@ template.replace(runWovenClassesTests())
   _: #@ template.replace(runSourceGenerationTests())
@@ -47,3 +47,4 @@ jobs:
     uses: ./.github/workflows/test-weaver.yml
     name: Test
   _: #@ template.replace(runTests("Code Coverage"))
+  _: #@ template.replace(cleanupBaas(["Code Coverage"]))
