+++ Tests/Realm.Tests/Sync/SyncTestHelpers.cs
@@ -149,7 +149,7 @@ namespace Realms.Tests.Sync
             return remainingArgs;
         }
 
-        public static (string[] RemainingArgs, AsyncFileLogger Logger) SetLoggerFromArgs(string[] args)
+        public static (string[] RemainingArgs, IDisposable Logger) SetLoggerFromArgs(string[] args)
         {
             var (extracted, remaining) = ArgumentHelper.ExtractArguments(args, "realmloglevel", "realmlogfile");
 
@@ -160,7 +160,7 @@ namespace Realms.Tests.Sync
                 Logger.LogLevel = logLevel;
             }
 
-            AsyncFileLogger logger = null;
+            Logger.AsyncFileLogger logger = null;
             if (extracted.TryGetValue("realmlogfile", out var logFile))
             {
                 if (!Process.GetCurrentProcess().ProcessName.ToLower().Contains("testhost"))
@@ -175,7 +175,7 @@ namespace Realms.Tests.Sync
                     TestHelpers.Output.WriteLine($"Setting async logger to file: {logFile}");
 
                     // We're running standalone (likely on CI), so we use the async logger
-                    Logger.Default = logger = new AsyncFileLogger(logFile);
+                    Logger.Default = logger = new Logger.AsyncFileLogger(logFile);
                 }
             }
 
@@ -221,64 +221,6 @@ namespace Realms.Tests.Sync
         {
             var app = _apps[appConfigType];
             return _baasClient.SetAutomaticRecoveryEnabled(app, enabled);
-        }
-
-        public class AsyncFileLogger : Logger, IDisposable
-        {
-            private readonly ConcurrentQueue<string> _queue = new();
-            private readonly string _filePath;
-            private readonly Encoding _encoding;
-            private readonly AutoResetEvent _hasNewItems = new(false);
-            private readonly AutoResetEvent _flush = new(false);
-            private readonly Task _runner;
-            private volatile bool _isFlushing;
-
-            public AsyncFileLogger(string filePath, Encoding encoding = null)
-            {
-                _filePath = filePath;
-                _encoding = encoding ?? Encoding.UTF8;
-                _runner = Task.Run(Run);
-            }
-
-            public void Dispose()
-            {
-                _isFlushing = true;
-                _flush.Set();
-                _runner.Wait();
-
-                _hasNewItems.Dispose();
-                _flush.Dispose();
-            }
-
-            protected override void LogImpl(LogLevel level, string message)
-            {
-                if (!_isFlushing)
-                {
-                    _queue.Enqueue(FormatLog(level, message));
-                    _hasNewItems.Set();
-                }
-            }
-
-            private void Run()
-            {
-                while (true)
-                {
-                    WaitHandle.WaitAny(new[] { _hasNewItems, _flush });
-
-                    var sb = new StringBuilder();
-                    while (_queue.TryDequeue(out var item))
-                    {
-                        sb.AppendLine(item);
-                    }
-
-                    System.IO.File.AppendAllText(_filePath, sb.ToString(), _encoding);
-
-                    if (_isFlushing)
-                    {
-                        return;
-                    }
-                }
-            }
         }
     }
 }
